Level scle to fit ui elements

+++ core/src/com/group25/game/Player.java
@@ -6,7 +6,7 @@ import java.awt.*;
 
 public class Player extends Creature{
 
-    public Player(float positionX, float positionY,int width, int height,int health, Sprite img, int entitySpeed){
+    public Player(float positionX, float positionY,int width, int height,int health, Sprite img, float entitySpeed){
         super(positionX, positionY, width, height, health, img, entitySpeed, new Rectangle((int)positionX,(int)positionY,100,100));
     }
